# openapi.yaml
openapi: 3.2.0
info:
  title: SDI - Servicio Distribuido de Inventario
  summary: API de gestion de stock para integracion entre modulos ERP.
  description: '

    API REST del modulo de Stock del sistema SDI.


    Permite administrar productos, consultar stock, crear reservas,

    confirmar ventas, liberar reservas y procesar eventos de compra.


    Esta API esta pensada para integracion entre sistemas internos:

    - Modulo de Ventas

    - Modulo de Compras

    - Modulo de Stock

    - Servicios administrativos internos


    La primera version no incluye seguridad formal. El consumo debe realizarse

    en ambiente controlado de red interna o infraestructura protegida.

    '
  termsOfService: https://docs.sdi.local/terminos-api
  contact:
    name: Equipo SDI - Sistema Distribuido de Inventario
    email: gxbeto@gmail.com
  license:
    name: Uso de la catedra Programacion de Aplicaiciones en Red - Profesor Alfonso
      Fernandez
  version: 0.1.0
servers:
- url: http://127.0.0.1:8000
  description: Ambiente local de desarrollo
- url: http://sdi-pruebas.example.com:6000
  description: Ambiente de pruebas / integracion
paths:
...
  /productos:
    post:
      tags:
      - productos
      summary: Crear producto e inicializar su stock en cero
      description: Registra un producto activo en el catalogo e inserta automaticamente
        su fila en stock_actual con cantidades en cero.
      operationId: crearProducto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductoCreate'
      responses:
        '201':
          description: Producto creado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductoCreated'
              example:
                producto_id: 1
                codigo: INF-006
                estado: CREADO
        '409':
          description: Ya existe un producto con el mismo codigo.
          content:
            application/json:
              example:
                estado: ERROR
                codigo: PRODUCTO_CODIGO_DUPLICADO
                mensaje: Ya existe un producto con el codigo INF-006.
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          $ref: '#/components/responses/ValidationErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
        default:
          $ref: '#/components/responses/DefaultErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedErrorResponse'
        '403':
          $ref: '#/components/responses/ForbiddenErrorResponse'
    get:
      tags:
      - productos
      summary: Listar productos con filtros opcionales
      description: Lista productos del catalogo. Permite filtrar por codigo exacto
        y por estado activo/inactivo.
      operationId: listarProductos
      parameters:
      - name: filtro
        in: query
        required: false
        schema:
          type: string
          minLength: 1
          maxLength: 50
        description: Código exacto del producto. Escriba cualquier código, por ejemplo
          INF-006.
      - name: activo
        in: query
        required: false
        schema:
          type: boolean
        description: Estado del producto. Use true para activos y false para inactivos.
      responses:
        '200':
          description: Lista de productos.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductoOut'
                title: Response Listar Productos Productos Get
              example:
              - producto_id: 1
                codigo: INF-006
                nombre: Teclado Satellite AK-910 USB / Negro
                descripcion: Teclado USB con cable, color negro
                categoria: Informatica
                unidad_medida: UNIDAD
                impuesto: 10
                activo: true
              - producto_id: 13
                codigo: CEL-001
                nombre: Celular Apple iPhone 17 256GB
                descripcion: Smartphone Apple, 256GB de almacenamiento
                categoria: Celulares
                unidad_medida: UNIDAD
                impuesto: 10
                activo: true
        '422':
          $ref: '#/components/responses/ValidationErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
        default:
          $ref: '#/components/responses/DefaultErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedErrorResponse'
        '403':
          $ref: '#/components/responses/ForbiddenErrorResponse'
  /productos/{producto_id}:
    get:
      tags:
      - productos
      summary: Consultar producto por identificador (consulta flexible agrupada)
      description: 'Devuelve una respuesta agrupada con los bloques solicitados en
        el parámetro `include`: producto, stock, precios y/o movimientos (separados
        por coma). Si no se informa include, se asume `producto`. El orden no afecta
        el resultado y los duplicados se ignoran. Un valor no permitido responde 422
        VALIDACION_ERROR. No reemplaza a GET /stock/{producto_id} ni a GET /stock/movimientos.'
      operationId: obtenerProducto
      parameters:
      - name: producto_id
        in: path
        required: true
        schema:
          type: integer
          title: Producto Id
          minimum: 1
          description: Debe ser un identificador entero positivo.
      - name: include
        in: query
        required: false
        schema:
          type: string
          description: 'Bloques a incluir, separados por coma. Valores permitidos:
            producto, stock, precios, movimientos. Ej.: include=producto,stock,precios'
          default: producto
          title: Include
        description: 'Bloques a incluir en la respuesta agrupada, separados por coma.
          Valores permitidos: producto, stock, precios, movimientos. Opcional; si
          se omite se asume ''producto''. El orden no afecta el resultado y los duplicados
          se ignoran. Ej.: include=producto,stock,precios.'
        style: form
        explode: false
      - name: limite_movimientos
        in: query
        required: false
        schema:
          type: integer
          maximum: 50
          minimum: 1
          description: 'Cantidad máxima de movimientos (solo aplica si include contiene
            ''movimientos''). Orden: más recientes primero.'
          default: 10
          title: Limite Movimientos
        description: 'Cantidad máxima de movimientos (solo aplica si include contiene
          ''movimientos''). Orden: más recientes primero.'
      responses:
        '200':
          description: Producto consultado. Respuesta agrupada según los bloques de
            include.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductoConsultaOut'
              examples:
                producto:
                  summary: include=producto
                  value:
                    producto:
                      producto_id: 1
                      codigo: INF-006
                      nombre: Teclado Satellite AK-910 USB / Negro
                      descripcion: Teclado USB con cable, color negro
                      categoria: Informatica
                      unidad_medida: UNIDAD
                      impuesto: 10
                      activo: true
                stock:
                  summary: include=stock
                  value:
                    stock:
                      producto_id: 1
                      cantidad_total: 10
                      cantidad_reservada: 2
                      cantidad_disponible: 8
                productoStockPrecios:
                  summary: include=producto,stock,precios
                  value:
                    producto:
                      producto_id: 1
                      codigo: INF-006
                      nombre: Teclado Satellite AK-910 USB / Negro
                      descripcion: Teclado USB con cable, color negro
                      categoria: Informatica
                      unidad_medida: UNIDAD
                      impuesto: 10
                      activo: true
                    stock:
                      producto_id: 1
                      cantidad_total: 10
                      cantidad_reservada: 2
                      cantidad_disponible: 8
                    precios:
                      precio_compra: 32000
                      precio_venta: 45760
                stockMovimientos:
                  summary: include=stock,movimientos&limite_movimientos=10
                  value:
                    stock:
                      producto_id: 1
                      cantidad_total: 10
                      cantidad_reservada: 2
                      cantidad_disponible: 8
                    movimientos:
                    - movimiento_id: 50
                      producto_id: 1
                      reserva_id: 101
                      tipo_movimiento: VENTA
                      origen: VENTAS
                      documento_ref: FV-001-000123
                      comprobante: STK-001-000001-6
                      cantidad: 2
                      stock_anterior: 10
                      stock_posterior: 8
                      observacion: Confirmación de venta desde reserva
                productoMovimientos:
                  summary: include=producto,movimientos&limite_movimientos=10
                  value:
                    producto:
                      producto_id: 1
                      codigo: INF-006
                      nombre: Teclado Satellite AK-910 USB / Negro
                      descripcion: Teclado USB con cable, color negro
                      categoria: Informatica
                      unidad_medida: UNIDAD
                      impuesto: 10
                      activo: true
                    movimientos:
                    - movimiento_id: 50
                      producto_id: 1
                      reserva_id: 101
                      tipo_movimiento: VENTA
                      origen: VENTAS
                      documento_ref: FV-001-000123
                      comprobante: STK-001-000001-6
                      cantidad: 2
                      stock_anterior: 10
                      stock_posterior: 8
                      observacion: Confirmación de venta desde reserva
        '404':
          description: El producto solicitado no existe.
          content:
            application/json:
              example:
                estado: ERROR
                codigo: PRODUCTO_NO_EXISTE
                mensaje: El producto 1 no existe.
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          $ref: '#/components/responses/ValidationErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
        default:
          $ref: '#/components/responses/DefaultErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedErrorResponse'
        '403':
          $ref: '#/components/responses/ForbiddenErrorResponse'
...
  /stock/reservas:
    post:
      tags:
      - stock
      summary: Crear reserva de stock para uno o mas productos
      description: 'Valida productos activos y stock disponible, bloquea las filas
        de stock involucradas, incrementa la cantidad reservada y registra movimientos
        de tipo RESERVA.


        La operación debe tratar reintentos equivalentes de forma idempotente. Clave
        idempotente: documento_ref.'
      operationId: crearReserva
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReservaCreate'
        required: true
      responses:
        '201':
          description: Reserva creada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservaCreateOut'
              example:
                estado: RESERVADO
                documento_ref: FV-001-000123
                comprobante: STK-001-000001-6
                reservas:
                - reserva_id: 101
                  producto_id: 1
                  cantidad_reservada: 2
                  cantidad_disponible: 8
        '400':
          description: El producto existe, pero no puede operar.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                estado: ERROR
                codigo: PRODUCTO_INACTIVO
                mensaje: El producto 1 esta inactivo.
        '404':
          description: El producto solicitado no existe.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                estado: ERROR
                codigo: PRODUCTO_NO_EXISTE
                mensaje: El producto 1 no existe.
        '409':
          description: Conflicto funcional, replay no coincidente, operación ya procesada
            o conflicto de concurrencia.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                stockInsuficiente:
                  summary: Stock insuficiente
                  value:
                    estado: ERROR
                    codigo: STOCK_INSUFICIENTE
                    mensaje: 'Stock insuficiente para el producto 1: solicitado 2,
                      disponible 1.'
                documentoYaTieneReserva:
                  summary: Documento ya reservado
                  value:
                    estado: ERROR
                    codigo: DOCUMENTO_YA_TIENE_RESERVA
                    mensaje: El documento FV-001-000123 ya posee una reserva registrada.
                replayNoCoincide:
                  summary: Replay con contenido diferente
                  value:
                    estado: ERROR
                    codigo: REPLAY_NO_COINCIDE
                    mensaje: El documento_ref ya existe, pero los ítems recibidos
                      no coinciden con la reserva original.
                conflictoConcurrencia:
                  summary: Conflicto de concurrencia
                  value:
                    estado: ERROR
                    codigo: CONFLICTO_CONCURRENCIA
                    mensaje: La reserva no pudo completarse porque el stock fue modificado
                      por otra operación concurrente.
        '422':
          $ref: '#/components/responses/ValidationErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
        default:
          $ref: '#/components/responses/DefaultErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedErrorResponse'
        '403':
          $ref: '#/components/responses/ForbiddenErrorResponse'
      x-idempotency:
        enabled: true
        keyField: documento_ref
        scope: Documento externo de venta o pedido que solicita la reserva.
        replayBehavior: Si llega el mismo documento_ref con los mismos ítems, no se
          duplica la reserva; se devuelve el estado ya registrado.
        payloadConflictBehavior: Si llega el mismo documento_ref con ítems diferentes,
          se rechaza con 409 y código REPLAY_NO_COINCIDE.
        sameRequestPolicy: No debe duplicar movimientos, reservas ni descuentos ante
          reintentos equivalentes.
        differentPayloadPolicy: Debe responder 409 si la clave idempotente ya existe
          con contenido diferente.
      x-concurrency:
        strategy: pessimistic-locking
        conflictStatus: 409
        conflictCode: CONFLICTO_CONCURRENCIA
        description: Las operaciones que afectan stock deben ejecutarse en transacción
          y proteger el saldo/reserva contra modificaciones concurrentes.
  /stock/reservas/{reserva_id}/confirmar:
    post:
      tags:
      - stock
      summary: Confirmar reserva y descontar stock definitivo
      description: 'Confirma una reserva pendiente, descuenta cantidad_total y cantidad_reservada,
        registra movimiento de VENTA y opcionalmente precio de venta.


        Operación de uso interno. Solo debe ser invocada por módulos internos del
        ecosistema SDI, principalmente Ventas o procesos administrativos autorizados.
        No forma parte del contrato público para integradores externos.


        La operación debe tratar reintentos equivalentes de forma idempotente. Clave
        idempotente: reserva_id.'
      operationId: confirmarReserva
      parameters:
      - name: reserva_id
        in: path
        required: true
        schema:
          type: integer
          title: Reserva Id
          minimum: 1
          description: Debe ser un identificador entero positivo.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmarReservaIn'
      responses:
        '200':
          description: Reserva confirmada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfirmarReservaOut'
              example:
                estado: CONFIRMADA
                reserva_id: 101
                producto_id: 1
                documento_ref: FV-001-000123
                comprobante: STK-001-000002-4
                cantidad_descontada: 2
                stock_total_actual: 8
                cantidad_reservada_actual: 0
                cantidad_disponible_actual: 8
        '404':
          description: La reserva solicitada no existe.
          content:
            application/json:
              example:
                estado: ERROR
                codigo: RESERVA_NO_EXISTE
                mensaje: La reserva 101 no existe.
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflicto funcional, replay no coincidente, operación ya procesada
            o conflicto de concurrencia.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                operacionYaProcesada:
                  summary: Confirmación repetida
                  value:
                    estado: ERROR
                    codigo: OPERACION_YA_PROCESADA
                    mensaje: La reserva 101 ya fue confirmada previamente.
                reservaYaLiberada:
                  summary: Reserva liberada previamente
                  value:
                    estado: ERROR
                    codigo: RESERVA_YA_LIBERADA
                    mensaje: La reserva 101 ya fue liberada y no puede confirmarse.
                conflictoConcurrencia:
                  summary: Conflicto de concurrencia
                  value:
                    estado: ERROR
                    codigo: CONFLICTO_CONCURRENCIA
                    mensaje: La reserva fue modificada por otra operación concurrente.
        '422':
          $ref: '#/components/responses/ValidationErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
        default:
          $ref: '#/components/responses/DefaultErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedErrorResponse'
        '403':
          $ref: '#/components/responses/ForbiddenErrorResponse'
      x-internal: true
      x-idempotency:
        enabled: true
        keyField: reserva_id
        scope: Confirmación interna de una reserva específica.
        replayBehavior: Si la reserva ya está CONFIRMADA, la repetición de la confirmación
          se trata como idempotente y devuelve el estado confirmado sin descontar
          stock nuevamente.
        payloadConflictBehavior: Si la reserva está LIBERADA o no corresponde a la
          operación, se rechaza con 409.
        sameRequestPolicy: No debe duplicar movimientos, reservas ni descuentos ante
          reintentos equivalentes.
        differentPayloadPolicy: Debe responder 409 si la clave idempotente ya existe
          con contenido diferente.
      x-concurrency:
        strategy: pessimistic-locking
        conflictStatus: 409
        conflictCode: CONFLICTO_CONCURRENCIA
        description: Las operaciones que afectan stock deben ejecutarse en transacción
          y proteger el saldo/reserva contra modificaciones concurrentes.
...
  /compras/eventos:
    post:
      tags:
      - compras
      summary: Procesar evento de compra e incrementar stock
      description: 'Procesa un evento de compra recibido por HTTP o reenviado desde
        RabbitMQ, evita duplicados por mensaje_id, incrementa stock y registra movimiento
        y precio de compra.


        La operación debe tratar reintentos equivalentes de forma idempotente. Clave
        idempotente: mensaje_id.'
      operationId: procesarEventoCompra
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompraEventoIn'
        required: true
      responses:
        '200':
          description: Evento de compra procesado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompraEventoOut'
              example:
                mensaje_id: CMP-2026-0001
                estado_procesamiento: PROCESADO
                producto_id: 1
                documento_ref: OC-001-000555
                comprobante: STK-001-000004-0
                stock_total_actual: 18
        '400':
          description: El producto existe, pero no puede operar.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                estado: ERROR
                codigo: PRODUCTO_INACTIVO
                mensaje: El producto 1 esta inactivo.
        '404':
          description: El producto solicitado no existe.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                estado: ERROR
                codigo: PRODUCTO_NO_EXISTE
                mensaje: El producto 1 no existe.
        '422':
          $ref: '#/components/responses/ValidationErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
        default:
          $ref: '#/components/responses/DefaultErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedErrorResponse'
        '403':
          $ref: '#/components/responses/ForbiddenErrorResponse'
        '409':
          description: Conflicto funcional, replay no coincidente, operación ya procesada
            o conflicto de concurrencia.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                mensajeDuplicado:
                  summary: Mensaje ya procesado
                  value:
                    estado: ERROR
                    codigo: MENSAJE_DUPLICADO
                    mensaje: El mensaje CMP-2026-0001 ya fue procesado previamente.
                replayNoCoincide:
                  summary: Replay con contenido diferente
                  value:
                    estado: ERROR
                    codigo: REPLAY_NO_COINCIDE
                    mensaje: El mensaje_id ya existe, pero el contenido recibido no
                      coincide con el procesamiento original.
                operacionEnProceso:
                  summary: Operación en proceso
                  value:
                    estado: ERROR
                    codigo: OPERACION_EN_PROCESO
                    mensaje: El evento de compra está siendo procesado. Reintente
                      la consulta o el envío más tarde.
      x-idempotency:
        enabled: true
        keyField: mensaje_id
        scope: Evento de compra recibido desde Compras o reenviado desde RabbitMQ.
        replayBehavior: Si llega el mismo mensaje_id con el mismo contenido, no se
          vuelve a incrementar stock; se devuelve el resultado ya procesado.
        payloadConflictBehavior: Si llega el mismo mensaje_id con contenido diferente,
          se rechaza con 409 y código REPLAY_NO_COINCIDE.
        sameRequestPolicy: No debe duplicar movimientos, reservas ni descuentos ante
          reintentos equivalentes.
        differentPayloadPolicy: Debe responder 409 si la clave idempotente ya existe
          con contenido diferente.
      x-concurrency:
        strategy: pessimistic-locking
        conflictStatus: 409
        conflictCode: CONFLICTO_CONCURRENCIA
        description: Las operaciones que afectan stock deben ejecutarse en transacción
          y proteger el saldo/reserva contra modificaciones concurrentes.
...
components:
  schemas:
...
    ErrorResponse:
      properties:
        estado:
          type: string
          enum:
          - ERROR
          title: Estado
          description: Estado fijo para respuestas de error.
          default: ERROR
        codigo:
          type: string
          enum:
          - VALIDACION_ERROR
          - PRODUCTO_NO_EXISTE
          - PRODUCTO_INACTIVO
          - PRODUCTO_CODIGO_DUPLICADO
          - STOCK_NO_INICIALIZADO
          - STOCK_INSUFICIENTE
          - RESERVA_NO_EXISTE
          - RESERVA_NO_PENDIENTE
          - RESERVA_NO_COINCIDE
          - NO_AUTENTICADO
          - ACCESO_DENEGADO
          - ERROR_INTERNO
          - ERROR_NO_CONTROLADO
          - MENSAJE_DUPLICADO
          - REPLAY_NO_COINCIDE
          - OPERACION_YA_PROCESADA
          - OPERACION_EN_PROCESO
          - CONFLICTO_CONCURRENCIA
          - DOCUMENTO_YA_PROCESADO
          - DOCUMENTO_YA_TIENE_RESERVA
          - RESERVA_YA_CONFIRMADA
          - RESERVA_YA_LIBERADA
          title: Codigo
          description: Código técnico de error devuelto por la API.
        mensaje:
          type: string
          title: Mensaje
          examples:
          - Descripcion clara del error.
      type: object
      required:
      - codigo
      - mensaje
      title: ErrorResponse
      example:
        codigo: CODIGO_ERROR
        estado: ERROR
        mensaje: Descripcion clara del error.
...
    ProductoCreate:
      properties:
        codigo:
          type: string
          maxLength: 50
          minLength: 1
          title: Codigo
          examples:
          - INF-006
        nombre:
          type: string
          maxLength: 150
          minLength: 1
          title: Nombre
          examples:
          - Teclado Satellite AK-910 USB / Negro
        descripcion:
          anyOf:
          - type: string
          - type: 'null'
          title: Descripcion
        categoria:
          anyOf:
          - type: string
          - type: 'null'
          title: Categoria
          examples:
          - Informatica
        unidad_medida:
          type: string
          enum:
          - UNIDAD
          - CAJA
          - KG
          - LITRO
          - METRO
          title: Unidad Medida
          description: Unidad de medida utilizada para registrar y mover stock.
          default: UNIDAD
        impuesto:
          type: number
          multipleOf: 0.01
          maximum: 100.0
          minimum: 0.0
          format: decimal
          title: Impuesto
          default: '0.00'
          examples:
          - 10
          example: 10
      type: object
      required:
      - codigo
      - nombre
      title: ProductoCreate
      example:
        categoria: Informatica
        codigo: INF-006
        descripcion: Teclado USB con cable, color negro
        impuesto: 10
        nombre: Teclado Satellite AK-910 USB / Negro
        unidad_medida: UNIDAD
    ProductoCreated:
      properties:
        producto_id:
          type: integer
          title: Producto Id
          minimum: 1
          description: Debe ser un identificador entero positivo.
        codigo:
          type: string
          title: Codigo
        estado:
          type: string
          enum:
          - CREADO
          title: Estado
          description: Estado devuelto cuando el producto fue creado.
          default: CREADO
      type: object
      required:
      - producto_id
      - codigo
      title: ProductoCreated
      example:
        codigo: INF-006
        estado: CREADO
        producto_id: 1
...
  responses:
    ValidationErrorResponse:
      description: El request no cumple el contrato OpenAPI o las validaciones declaradas.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            estado: ERROR
            codigo: VALIDACION_ERROR
            mensaje: 'campo: descripción clara de la validación incumplida.'
    UnauthorizedErrorResponse:
      description: No autenticado. Falta credencial válida o el mecanismo de autenticación
        rechazó la solicitud.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            estado: ERROR
            codigo: NO_AUTENTICADO
            mensaje: No se pudo autenticar la solicitud.
    ForbiddenErrorResponse:
      description: Acceso denegado. El consumidor no tiene permiso para ejecutar la
        operación solicitada.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            estado: ERROR
            codigo: ACCESO_DENEGADO
            mensaje: No tiene permisos para ejecutar esta operación.
    InternalServerErrorResponse:
      description: Error interno controlado del servicio.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            estado: ERROR
            codigo: ERROR_INTERNO
            mensaje: Ocurrió un error interno al procesar la operación.
    DefaultErrorResponse:
      description: Error no controlado o no clasificado por el contrato.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            estado: ERROR
            codigo: ERROR_NO_CONTROLADO
            mensaje: Ocurrió un error no clasificado por el contrato.
...